Altered README file completely
I made many changes to the README file. I removed a section that talked about the languages I wanted to learn in the current year because it's not a point of focus anymore. I changed the format of the file and added a personal motto at the bottom of the file.